Added a button that creates a Stripe Checkout session with an OrderDetails object, a test page can be seen in RedirectToPaymentForm
